Split into "main" and "suppress" components
