+++ addons/main/CfgMods.hpp
@@ -2,7 +2,7 @@ class CfgMods {
     class PREFIX {
         dir = "@SimpleSuppress";
         name = "Simple Suppress";
-        actionName = "Github";
+        actionName = "GitHub";
         action = CSTRING(URL);
         description = "Arma 3 mod providing simple suppression effect for players.";
     };
