--- addons/main/functions/script_component.hpp
@@ -1,1 +1,0 @@
-#include "\z\simplesuppress\addons\main\script_component.hpp"
