+++ addons/suppress/$PBOPREFIX$
@@ -1,0 +1,1 @@
+z\simplesuppress\addons\suppress
