+++ addons/suppress/functions/fnc_handlePlayerChanged.sqf
