+++ addons/suppress/functions/fnc_handleSuppressed.sqf
