+++ addons/suppress/functions/fnc_resetEffects.sqf
