+++ addons/suppress/functions/script_component.hpp
@@ -1,0 +1,1 @@
+#include "..\script_component.hpp"
