+++ addons/suppress/script_component.hpp
@@ -1,0 +1,11 @@
+#define COMPONENT suppress
+#define COMPONENT_BEAUTIFIED Suppress
+#include "\z\simplesuppress\addons\main\script_mod.hpp"
+
+// #define DEBUG_MODE_FULL
+// #define DISABLE_COMPILE_CACHE
+// #define ENABLE_PERFORMANCE_COUNTERS
+
+#include "\z\simplesuppress\addons\main\script_macros.hpp"
+
+#define IDC_OVERLAY 1202
